==> 2- Data Storage.ipynb <==
# 1- Tabelas Auxiliares  

Serão utilizadas para interpretar os códigos da base de dados e, por isso, devem ser tratadas no primeiro passo.

### Distribuição Territorial Brasileira - IBGE  

Devido a uma demonstração de interesse por parte de uma empresa importadora de fertilizantes sediada em Pouso Alegre/MG, esse projeto visa prever as importações deste município e de todos os demais na mesma região geográfica.  

O arquivo coletado e armazenado no endereço a seguir será descompactado e filtrado, deixando passar somente as informações referentes a região de Pouso Alegre/MG.

* '~/git/BD/IBGE/DTB_2022.zip'

Ao procurar o nome "Pouso Alegre" dentre os tipos de sub-divisão geograficas, se confirma que este município é classificado como centro de uma "Região Geográfica **Intermediária**" do estado de Minas Gerais. As regiões deste tipo contém nelas uma segunda sub-divisão, denominada "Região Geográfica **Imediata**" onde, por fim, se encontra a última e menor subdivisão, denominada Município.

Ao final, o arquivo será salvo em :

* 'git/Importacao-por-SH-e-Municipio/temp/i_f_r_pa.csv'

%run '~/git/Bibliotecas.ipynb'

import zipfile
import io
import os
import pandas as pd

caminho_arquivo_zip = '~/git/BD/IBGE/DTB_2022.zip'
string_procurada = 'MUNICIPIO'

# Expandir o caminho do arquivo
caminho_arquivo_zip = os.path.expanduser(caminho_arquivo_zip)

# Variável para armazenar o conteúdo do arquivo XLS
m = None

# Descompactar o arquivo em memória
with zipfile.ZipFile(caminho_arquivo_zip, 'r') as zip_ref:
    # Obter uma lista de nomes de arquivos no zip
    nomes_arquivos = zip_ref.namelist()

    for nome_arquivo in nomes_arquivos:
        # Verificar se a string 'MUNICIPIO' está presente no nome do arquivo e se o arquivo é um XLS
        if string_procurada.upper() in nome_arquivo.upper() and nome_arquivo.lower().endswith('.xls'):
            # Ler o conteúdo do arquivo no zip
            conteudo_arquivo = zip_ref.read(nome_arquivo)

            # Utilizar o pandas para ler o arquivo XLS
            m = pd.read_excel(io.BytesIO(conteudo_arquivo), skiprows=6)

            # Exibir as primeiras linhas do DataFrame (opcional)
            display(m.head())

            # Aqui você pode fazer operações adicionais com o conteúdo do arquivo, se necessário

import os

folder_path = '/home/andre301267/git/Importacao-por-SH-e-Municipio/temp'

# Check if the folder already exists
if not os.path.exists(folder_path):
    # Create the folder
    os.makedirs(folder_path)
    print(f"Folder '{folder_path}' created successfully.")
else:
    print(f"Folder '{folder_path}' already exists.")


# Filtra deixando passar somente municíppios da região de Pouso Alegere
m_r_pa=m[
    (m['Nome Região Geográfica Intermediária'].astype(str).str.contains('Pouso Alegre'))
    &
    (m['Nome_UF']=='Minas Gerais')
]

# seleciona somente a coluna com o nome dos municípios
m_r_pa=m_r_pa[m_r_pa.columns[[12]].tolist()]
# renomeia a coluna em busca de uniformidade com outros arquivos 
m_r_pa=m_r_pa.rename(columns={'Nome_Município':'Município'}).reset_index(drop=True)

# Salva 
m_r_pa.to_csv('~/git/comex_sh_mun/temp/m_r_pa.csv',index=False)

# Imprime
print('\nMunicípios da Região de Pouso Alegre:')
display(m_r_pa)
print('\nFonte: IBGE')

# 2- Importação de Fertilizantes por Município  

No capítulo anterior, a base de dados com importações por Município, denominada IMP_COMPLETA_MUN.zip, e as tabelas complementares, foram baixadas e salvas localmente na pasta:  

* '~/git/BD/Comexstat/'

Descompacta o arquivo e salva localmente em pasta específica

import zipfile
import os

caminho_arquivo_zip = '/home/andre301267/git/BD/Comexstat/IMP_COMPLETA_MUN.zip'
diretorio_destino = '/home/andre301267/git/BD/Comexstat/extracted_files/'

# Crie o diretório de destino, se não existir
os.makedirs(diretorio_destino, exist_ok=True)

# Abra o arquivo ZIP e extraia para o diretório de destino
with zipfile.ZipFile(caminho_arquivo_zip, 'r') as arquivo_zip:
    arquivo_zip.extractall(diretorio_destino)

print(f'Arquivos extraídos para: {diretorio_destino}')

Lê in chunk e Imprime as primeiras 5 linhas do primeiro chunk, para verificar a estrutura do arquivo

reader=pd.read_csv('~/git/BD/Comexstat/extracted_files/IMP_COMPLETA_MUN.csv', chunksize=5, encoding='Latin-1',sep=';')

print(next(reader))

Nessa tabela acima, o produto é identificado na coluna SH4. Para interpretar esse código numérico e filtrar somente os fertilizantes, será necessário funidir a base de dados com a tabela auxiliar "NCM_SH.csv ", fornecida pelo site Comexstat.  

Vamos então abrir e visualizar a tabela "NCM_SH.csv":

ncm_sh=pd.read_csv('~/git/BD/Comexstat/NCM_SH.csv',encoding='Latin-1',sep=';')
ncm_sh.head(1)

ncm_sh.info()

O arquivo original o detalhamento de produto é SH4. Tentaremos simplificar buscando localizar os fertilizantes em um grupamento de produto mais amplo, o SH2

ncm_sh=ncm_sh[ncm_sh.columns[[4,5,8,9]]].drop_duplicates()
ncm_sh.head()

Entre a base e dados e a tabela auxiliar, podemos notar que a colunas referentes ao código SH4 tem nomes diferentes. A seguir será alterado um dos nomes para iguala-los entre as tabelas.

ncm_sh=ncm_sh.rename(columns={'CO_SH4':'SH4'})

Na coluna com o nome do código SH2 em português, "NO_SH2_POR":

sh2_desejado=ncm_sh[ncm_sh.NO_SH2_POR.str.contains('fertilizantes')].CO_SH2.unique()
sh2_desejado

Como só foi localizado um único código SH2 contendo o termo 'fertilizantes', salvaremos ele em variável para na sequencia filtrarmos a base de dados.

sh2_desejado=sh2_desejado[0]
sh2_desejado

O arquivo da base de dados é muito grande para ser aberto de uma só vez. Vamos então ler o arquivo em chuncks, fundi-lo (merge) com a tabela auxiliar ncm_sh, e filtrar cada chunk que contiver na coluna "CO_SH2" o valor contido  na variável "sh2_desejado".

%%time
reader=pd.read_csv('~/git/BD/Comexstat/extracted_files/IMP_COMPLETA_MUN.csv', chunksize=100_000, encoding='Latin-1',sep=';')
i_f_m=pd.DataFrame()
for chunk in reader:
    chunk=chunk.merge(ncm_sh)
    chunk=chunk[chunk.CO_SH2==sh2_desejado]
    i_f_m=pd.concat([i_f_m,chunk])
i_f_m.head(1)

i_f_m.shape

i_f_m.duplicated(subset=['CO_ANO','CO_MES','SH4','CO_MUN','CO_PAIS']).sum()

Continuando a filtragem, dexando dessa vez passar somente as linhas onde a coluna SG_UF_MUN for igual a 'MG'

i_f_m=i_f_m[i_f_m.SG_UF_MUN=='MG']

Processando a última filtragem, agora deixando passar somente as linhas que correspndem aos municípios da região de Pouso Alegre. Para isso será utilizada a tabela auxiliar contendo os códigos e os nomes dos municípios, UF_MUN.csv

uf_mun=pd.read_csv('~/git/BD/Comexstat/UF_MUN.csv',encoding='Latin-1',sep=';')

Será modificado o nome deu uma das colunas de forma a igualar com a base de dados

uf_mun=uf_mun.rename(columns={'CO_MUN_GEO':'CO_MUN'})

Crrega a tabela  com os municípios da região de Pouso Alegre (m_r_pa)

m_r_pa=pd.read_csv('~/git/comex_sh_mun/temp/m_r_pa.csv').rename(columns={'Município':'NO_MUN_MIN'})

Procede a filtragem através da fundição das tabelas

i_f_m.duplicated().sum()

i_f_m_r_pa=uf_mun.merge(m_r_pa).merge(i_f_m)

i_f_m_r_pa.duplicated().sum()

# Salva a tabela resultante
i_f_m_r_pa.to_csv('~/git/comex_sh_mun/temp/i_f_m_r_pa.csv',index=False)

i_f_m_r_pa.head(1)

i_f_m_r_pa.duplicated().sum()

Por fim, apaga a pasta onde o arquivo original e descompactado foi salvo

import shutil

diretorio_destino = '/home/andre301267/git/BD/Comexstat/extracted_files/'

# Excluir o diretório e todo o seu conteúdo
shutil.rmtree(diretorio_destino)

print(f'Diretório {diretorio_destino} e seu conteúdo foram excluídos.')


# 3- Exportação de Café  

No capítulo anterior, a base de dados de exportação por Município, denominada EXP_COMPLETA_MUN.zip, e as tabelas complementares, foram baixadas e salvas localmente na pasta:  

* '~/git/BD/Comexstat/'

import zipfile
import io
import os
import pandas as pd

Descompacta o arquivo e salva localmente em pasta específica

import zipfile
import os

caminho_arquivo_zip = '/home/andre301267/git/BD/Comexstat/EXP_COMPLETA_MUN.zip'
diretorio_destino = '/home/andre301267/git/BD/Comexstat/extracted_files/'

# Crie o diretório de destino, se não existir
os.makedirs(diretorio_destino, exist_ok=True)

# Abra o arquivo ZIP e extraia para o diretório de destino
with zipfile.ZipFile(caminho_arquivo_zip, 'r') as arquivo_zip:
    arquivo_zip.extractall(diretorio_destino)

print(f'Arquivos extraídos para: {diretorio_destino}')

Lê in chunk e Imprime as primeiras 5 linhas do primeiro chunk, para verificar a estrutura do arquivo

reader=pd.read_csv('~/git/BD/Comexstat/extracted_files/EXP_COMPLETA_MUN.csv', chunksize=5, encoding='Latin-1',sep=';')

print(next(reader))

A tabela auxiliar contendo os códigos e os nomes dos municípios, UF_MUN.csv

uf_mun=pd.read_csv('~/git/BD/Comexstat/UF_MUN.csv',encoding='Latin-1',sep=';')
uf_mun.head(1)

Será modificado o nome de uma das colunas de forma a igualar com a base de dados

uf_mun=uf_mun.rename(columns={'CO_MUN_GEO':'CO_MUN'})

Carrega a tabela (que foi salva no item 1) com os municípios da região de Pouso Alegre (m_r_pa), para ser usada na filtragem os registros de exportação.

m_r_pa=pd.read_csv('~/git/comex_sh_mun/temp/m_r_pa.csv').rename(columns={'Município':'NO_MUN_MIN'})

co_mun_r_pa=uf_mun.merge(m_r_pa)

len(co_mun_r_pa[['CO_MUN']])

len(co_mun_r_pa[['CO_MUN']].drop_duplicates())

%%time
reader=pd.read_csv('~/git/BD/Comexstat/extracted_files/EXP_COMPLETA_MUN.csv', chunksize=10_000, encoding='Latin-1',sep=';')
e_m_r_pa=pd.DataFrame()
for chunk in reader:
    # Filtra cada chunk , deixando passar somente os registros dos municípios da região de Pouso Alegre
    chunk=chunk.merge(co_mun_r_pa)
    e_m_r_pa=pd.concat([e_m_r_pa,chunk])
e_m_r_pa.head(1)

Agora identificaremos qual produto agrícola é o mais relevante para a região foco, para que ele seja utilizado como variável multipla na previsão dos fertilizantes.  

Antes, em nome da simplicidade, o plano é prever os produtos utilizando o código de categora mais abrangente disponível, que é o código SH2, que é obtido através da fusão da base de dados com a tabela auxiliar, NCM_SH.csv, fornecida pelo site Comexstat.

ncm_sh=pd.read_csv('~/git/BD/Comexstat/NCM_SH.csv',encoding='Latin-1',sep=';').rename(columns={'CO_SH4':'SH4'})
ncm_sh.head(1)

Descarta detalhamentos maiores (SH6) e condensa a tabela. Sem esse procedimento, nas operações merge a frente, duplicatas seriam criadas.

ncm_sh=ncm_sh[['SH4', 'NO_SH4_POR', 'CO_SH2', 'NO_SH2_POR']].drop_duplicates()

len(e_m_r_pa)

e_m_r_pa=e_m_r_pa.merge(ncm_sh)

len(e_m_r_pa)

e_m_r_pa.groupby(by=['CO_SH2','NO_SH2_POR']).VL_FOB.sum().sort_values(ascending=False).reset_index().head()

e_c_m_r_pa=e_m_r_pa[e_m_r_pa.CO_SH2==9]

e_c_m_r_pa.head(1)

Enfim, salvamos o arquivo, que esta pronto para uso.

e_c_m_r_pa.to_csv('~/git/comex_sh_mun/temp/e_c_m_r_pa.csv',index=False)

Apaga arquivo zip expandido

import shutil
import os

# Diretório a ser apagado
directory_to_delete = '~/git/BD/Comexstat/extracted_files/'

# Expandir o caminho do usuário
directory_to_delete = os.path.expanduser(directory_to_delete)

# Verificar se o diretório existe antes de apagar
if os.path.exists(directory_to_delete):
    # Apagar o diretório e todo o seu conteúdo
    shutil.rmtree(directory_to_delete)
    print(f'Diretório {directory_to_delete} apagado com sucesso.')
else:
    print(f'O diretório {directory_to_delete} não existe.')

# 4- Volumes de Chuva  

No capítulo anterior, a base de dados sobre chuvas, que é representada por um arquivo zip por ano de coleta, foi baixadas e salvas localmente na pasta:  

* '~/git/BD/Inmet/'

Extrai um dos arquivos zip para reconhecer o que esta no inteior

%%time
import zipfile
import os

# Caminho do arquivo ZIP
caminho_arquivo_zip = os.path.expanduser('~/git/BD/Inmet/2020.zip')

# Expandir o arquivo ZIP para um diretório temporário
diretorio_temporario = '/home/andre301267/git/BD/Inmet/extracted_files/'
with zipfile.ZipFile(caminho_arquivo_zip, 'r') as zip_ref:
    zip_ref.extractall(diretorio_temporario)

# Listar os arquivos no diretório temporário
arquivos_temporarios = os.listdir(diretorio_temporario)
'''
# Exibir os arquivos
for arquivo_temporario in arquivos_temporarios:
    print(arquivo_temporario)'''

Limpa diretorio de extração

%%time
# Limpar o diretório temporário (opcional)
for arquivo_temporario in arquivos_temporarios:
    caminho_completo = os.path.join(diretorio_temporario, arquivo_temporario)
    os.remove(caminho_completo)

# Remover o diretório temporário
os.rmdir(diretorio_temporario)

Extrai todos os arquivos zip

%%time
import zipfile
import os

# Diretório de origem dos arquivos ZIP
diretorio_origem = os.path.expanduser('~/git/BD/Inmet/')

# Diretório de destino para os arquivos extraídos
diretorio_destino = '/home/andre301267/git/BD/Inmet/extracted_files/'

# Garante que o diretório de destino existe
os.makedirs(diretorio_destino, exist_ok=True)

# Lista de arquivos ZIP no diretório de origem
arquivos_zip = [arquivo for arquivo in os.listdir(diretorio_origem) if arquivo.endswith('.zip')]

# Itera sobre os arquivos ZIP e extrai para o diretório de destino
for arquivo_zip in arquivos_zip:
    caminho_completo_zip = os.path.join(diretorio_origem, arquivo_zip)

    with zipfile.ZipFile(caminho_completo_zip, 'r') as zip_ref:
        # Extrai os arquivos para o diretório de destino
        zip_ref.extractall(diretorio_destino)

print("Extração concluída.")


Move os arquivos extraídos para a pasta raíz

import os
import shutil

diretorio_origem = '/home/andre301267/git/BD/Inmet/extracted_files/'

# Iterar sobre todas as subpastas
for pasta, subpastas, arquivos in os.walk(diretorio_origem):
    # Iterar sobre os arquivos em cada subpasta
    for arquivo in arquivos:
        # Construir os caminhos completos de origem e destino
        caminho_origem = os.path.join(pasta, arquivo)
        caminho_destino = os.path.join(diretorio_origem, arquivo)

        # Mover o arquivo para a raiz do diretório
        shutil.move(caminho_origem, caminho_destino)

print("Movimentação concluída.")


Elimina as subpastas que ficaram vazias

import os

diretorio_origem = '/home/andre301267/git/BD/Inmet/extracted_files/'

# Iterar sobre todas as subpastas em ordem reversa
for pasta, subpastas, arquivos in os.walk(diretorio_origem, topdown=False):
    for subpasta in subpastas:
        # Construir o caminho completo da subpasta
        caminho_subpasta = os.path.join(pasta, subpasta)

        # Verificar se a subpasta está vazia
        if not os.listdir(caminho_subpasta):
            # Remover a subpasta vazia
            os.rmdir(caminho_subpasta)

print("Subpastas vazias removidas.")


Remove todos os arquivos descompatados que não se refiram ao estado de MG

import os

diretorio_origem = '/home/andre301267/git/BD/Inmet/extracted_files/'

# Iterar sobre todos os arquivos na raiz do diretório
for arquivo in os.listdir(diretorio_origem):
    # Construir o caminho completo do arquivo
    caminho_arquivo = os.path.join(diretorio_origem, arquivo)

    # Verificar se o arquivo não contém "_MG_" no nome
    if "_MG_" not in arquivo:
        # Remover o arquivo
        os.remove(caminho_arquivo)
        #print(f"Arquivo removido: {caminho_arquivo}")

print("Remoção de arquivos concluída.")


Carrega arquivo contendo lista de municípios da região de Pouso Alegre

import pandas as pd
m_r_pa=pd.read_csv('~/git/comex_sh_mun/temp/m_r_pa.csv')
m_r_pa.Município=m_r_pa.Município.str.upper().str.strip()
m_r_pa_l=m_r_pa.Município.tolist()

Remove os arquivos que não contiverem no nome um nome de um dos município da região de Pouso Alegre

import os

diretorio_origem = '/home/andre301267/git/BD/Inmet/extracted_files/'

# Iterar sobre todos os arquivos na raiz do diretório
for arquivo in os.listdir(diretorio_origem):
    # Construir o caminho completo do arquivo
    caminho_arquivo = os.path.join(diretorio_origem, arquivo)

    # Verificar se o arquivo não contém nenhuma das strings da lista
    if all(string not in arquivo for string in m_r_pa_l):
        # Remover o arquivo
        os.remove(caminho_arquivo)
        print(f"Arquivo removido: {caminho_arquivo}")

# print("Remoção de arquivos concluída.")

Divide as tabelas extraídas em duas, concatenando essas partes em uma planilha cada. Uma parte contedo a identificação das estações meteorológicas, outra parte contendo os dados climatológicos.

import os
import pandas as pd

# Diretório dos arquivos
diretorio = '/home/andre301267/git/BD/Inmet/extracted_files/'

# Lista para armazenar os DataFrames
df_parte1 = pd.DataFrame()
df_parte2 = pd.DataFrame()

# Iterar sobre os arquivos no diretório
for arquivo in os.listdir(diretorio):
    if arquivo.endswith('.CSV'):  # Certifique-se de que estamos lidando apenas com arquivos de texto
        caminho_arquivo = os.path.join(diretorio, arquivo)

        # Ler as primeiras 8 linhas
        parte1 = pd.read_csv(caminho_arquivo, nrows=7,sep=';',encoding='Latin-1')
       # print(parte1)

        # Ler a partir da linha 9
        parte2 = pd.read_csv(caminho_arquivo, skiprows=8,sep=';',encoding='Latin-1').assign(Estação=parte1.iloc[1,1])

        # Adicionar DataFrames às listas
        df_parte1=pd.concat([df_parte1,parte1],axis=1)
        df_parte2=pd.concat([df_parte2,parte2])

# Formata arquivo contendo ID das estações
df_parte1=df_parte1.transpose().drop(columns=[5,6]).drop_duplicates(subset=[1]).drop_duplicates(subset=[2])\
.reset_index(drop=True)

df_parte1.columns=df_parte1.iloc[0]

df_parte1=df_parte1.iloc[1:]


# Exibir os DataFrames
print("DataFrame Parte 1:")
display(df_parte1)

print("\nDataFrame Parte 2:")
display(df_parte2)

Remove o diretório de extração

import os

diretorio = '/home/andre301267/git/BD/Inmet/extracted_files/'

# Obtenha a versão absoluta do diretório
diretorio_absoluto = os.path.expanduser(diretorio)

# Liste todos os arquivos no diretório
arquivos = os.listdir(diretorio_absoluto)

# Itere sobre os arquivos e remova-os
for arquivo in arquivos:
    caminho_completo = os.path.join(diretorio_absoluto, arquivo)
    if os.path.isfile(caminho_completo):
        os.remove(caminho_completo)

print(f'Todos os arquivos em {diretorio} foram removidos.')

# Remover o diretório temporário
os.rmdir(diretorio)

Salva arquivos

df_parte1.to_csv('~/git/comex_sh_mun/temp/estações_meteo.csv',index=False)
df_parte2.to_csv('~/git/comex_sh_mun/temp/chuvas_m_r_pa.csv',index=False)

df_parte2['PRECIPITAÇÃO TOTAL, HORÁRIO (mm)'].sort_values().unique()
