==> 4- Analysys.ipynb <==
%run  '~/git/Bibliotecas.ipynb'

Cria função que procederá teste em todas as tabelas prreparadas para isso no capítulo anterior.  
Os testes serão executados utilizando todos os dados anteriores ao ano a ser testado. Os anos a serem testados são os últimos dez anos da tabela.

def test_rf(df, duração):
    # Cria df vazia para acumular os resultados dos testes
    q=pd.DataFrame()
    # Elimina dos testes o último ano da tabela, ano que esta em curso e é o objetivo da predição final.
    df_test=df[df.CO_ANO!=df.CO_ANO.max()]
    # Inicia loop para proceder teste nos últimos 10 anos 
    for a in range(df_test.CO_ANO.max()-duração,df_test.CO_ANO.max()+1):
        # Elimina anos posteriores ao ano de teste
        df_t=df_test[df_test.CO_ANO<=a]
        # Determina o período para treino, que são todos os anos menos o último
        train=df_t[df_t.CO_ANO!=df_t.CO_ANO.max()]
        # Determina o ano de teste, que é o último dsponível
        test=df_t[df_t.CO_ANO==df_t.CO_ANO.max()]
        # Determina X e y de treino
        X_train=train.drop(columns='KG_f')
        y_train=train.KG_f
        # Determina X e y de teste
        X_test=test.drop(columns='KG_f')
        y_test=test.KG_f
        # Fit o  regressor nos dados de teste
        RFR.fit(X_train,y_train)
        # Faz a previsão
        test_pred=test.assign(KG_f_pred=RFR.predict(X_test))
        #display(test_pred)
        
        # Cria tabela com resultado do  teste 
        w=pd.DataFrame({'Ano':[a],'KG_f':test_pred.KG_f,'KG_pred_f':test_pred.KG_f_pred})
        # Armazena os resultados
        q=pd.concat([q,w])
    '''    # Imprime o resultado do teste
        print(w)'''

    # Imprime os resultados gerais
    print(round(metrics.mean_absolute_percentage_error(q.KG_f,q.KG_pred_f),2))
    

i=pd.read_csv('~/git/comex_sh_mun/temp/i.csv')
i=i[i.CO_ANO!=i.CO_ANO.min()]
i_single=i[['CO_ANO','KG_f']]

i_wide=pd.read_csv('~/git/comex_sh_mun/temp/i_wide.csv')

i_e=pd.read_csv('~/git/comex_sh_mun/temp/i_e.csv')
i_e=i_e[i_e.CO_ANO!=i_e.CO_ANO.min()]

i_e_wide=pd.read_csv('~/git/comex_sh_mun/temp/i_e_wide.csv')
i_e_wide=i_e_wide[i_e_wide.CO_ANO!=i_e_wide.CO_ANO.min()]

i_e_c=pd.read_csv('~/git/comex_sh_mun/temp/i_e_c.csv')

i_e_c_wide=pd.read_csv('~/git/comex_sh_mun/temp/i_e_c_wide.csv')

test_rf(i_single,5), test_rf(i_single,2)

test_rf(i,5)

test_rf(i_wide,5)

test_rf(i,2), test_rf(i_wide,2)

test_rf(i_e,5), test_rf(i_e_wide,5)

test_rf(i_e,2) , test_rf(i_e_wide,2)

test_rf(i_e_c,5) ,test_rf(i_e_c_wide,5)

test_rf(i_e_c,5)

test_rf(i_e_c_wide,10)
