==> 6-Agrega Múltiplas Variáveis.ipynb <==
%run '~/git/Bibliotecas.ipynb'

print('\n\nFunde as tabelas de importações de fertilizantes, exportação de café e\n\
chuvas na região. Insere colunas referentes aos agrpamentos de 12 meses, casa um iniciado em um mes do ano.\n')

i_f_r_pa=pd.read_csv('~/git/Importacao-por-SH-e-Municipio/temp/i_f_r_pa.csv')
i_f_r_pa.head(5)

i_f_r_pa=i_f_r_pa\
.sort_values(by=['Ano','Mês'])\
.rename(columns={'Quilograma Líquido':'KT_f',\
                 'Valor FOB (US$)':'miFOB_f'})
i_f_r_pa[['KT_f','miFOB_f']]=i_f_r_pa[['KT_f','miFOB_f']]/10**6
i_f_r_pa['date']=pd.to_datetime(i_f_r_pa.Ano.astype(str)+'/'+i_f_r_pa.Mês.astype(str))
d_i=i_f_r_pa[['date']].iloc[0][0]
d_f=i_f_r_pa[['date']].iloc[-1][0]
i_f_r_pa[['Ano','Mês']]=i_f_r_pa[['Ano','Mês']].astype('category')
i_f_r_pa=i_f_r_pa.groupby(by=['Ano','Mês'])[['KT_f','miFOB_f']].sum().reset_index()
i_f_r_pa['date']=pd.to_datetime(i_f_r_pa.Ano.astype(str)+'/'+i_f_r_pa['Mês'].astype(str))
i_f_r_pa=i_f_r_pa[(i_f_r_pa.date>=d_i)&(i_f_r_pa.date<=d_f)]
i_f_r_pa

e_c_r_pa=pd.read_csv('~/git/Importacao-por-SH-e-Municipio/temp/e_c_r_pa.csv')
e_c_r_pa=e_c_r_pa\
.sort_values(by=['Ano','Mês'])\
.rename(columns={'Quilograma Líquido':'KT_c',\
                 'Valor FOB (US$)':'miFOB_c'})
e_c_r_pa[['KT_c','miFOB_c']]=e_c_r_pa[['KT_c','miFOB_c']]/10**6
e_c_r_pa['date']=pd.to_datetime(e_c_r_pa.Ano.astype(str)+'/'+e_c_r_pa.Mês.astype(str))
d_i=e_c_r_pa[['date']].iloc[0][0]
d_c=e_c_r_pa[['date']].iloc[-1][0]
e_c_r_pa[['Ano','Mês']]=e_c_r_pa[['Ano','Mês']].astype('category')
e_c_r_pa=e_c_r_pa.groupby(by=['Ano','Mês'])[['KT_c','miFOB_c']].sum().reset_index()
e_c_r_pa['date']=pd.to_datetime(e_c_r_pa.Ano.astype(str)+'/'+e_c_r_pa['Mês'].astype(str))
e_c_r_pa=e_c_r_pa[(e_c_r_pa.date>=d_i)&(e_c_r_pa.date<=d_c)]
e_c_r_pa

chuvas_r_pa=pd.read_csv('~/git/Importacao-por-SH-e-Municipio/temp/chuvas_r_pa.csv')
chuvas_r_pa=chuvas_r_pa\
.sort_values(by=['Ano','Mês'])


chuvas_r_pa['date']=pd.to_datetime(chuvas_r_pa.Ano.astype(str)+'/'+chuvas_r_pa.Mês.astype(str))

d_i=chuvas_r_pa[['date']].iloc[0][0]
d_f=chuvas_r_pa[['date']].iloc[-1][0]

chuvas_r_pa[['Ano','Mês']]=chuvas_r_pa[['Ano','Mês']].astype('category')
chuvas_r_pa=chuvas_r_pa.groupby(by=['Ano','Mês'])[['mm']].sum().reset_index()
chuvas_r_pa['date']=pd.to_datetime(chuvas_r_pa.Ano.astype(str)+'/'+chuvas_r_pa['Mês'].astype(str))
chuvas_r_pa=chuvas_r_pa[(chuvas_r_pa.date>=d_i)&(chuvas_r_pa.date<=d_f)]
chuvas_r_pa

r_pa=i_f_r_pa.merge(e_c_r_pa).merge(chuvas_r_pa)
r_pa=r_pa.rename(columns={'Ano':'Ano1'}).drop(columns='date')
r_pa

r_pa.isna().sum()

cal=r_pa[['Ano1','Mês']].drop_duplicates()
cal

cal['Ano2']=cal.Ano1
cal.loc[cal.Mês.isin([1]),'Ano2']=cal.Ano1-1

cal['Ano3']=cal.Ano1
cal.loc[cal.Mês.isin([1,2]),'Ano3']=cal.Ano1-1

cal['Ano4']=cal.Ano1
cal.loc[cal.Mês.isin([1,2,3]),'Ano4']=cal.Ano1-1

cal['Ano5']=cal.Ano1
cal.loc[cal.Mês.isin([1,2,3,4]),'Ano5']=cal.Ano1-1

cal['Ano6']=cal.Ano1
cal.loc[cal.Mês.isin([1,2,3,4,5]),'Ano6']=cal.Ano1-1

cal['Ano7']=cal.Ano1
cal.loc[cal.Mês.isin([1,2,3,4,5,6]),'Ano7']=cal.Ano1-1

cal['Ano8']=cal.Ano1
cal.loc[cal.Mês.isin([1,2,3,4,5,6,7]),'Ano8']=cal.Ano1-1

cal['Ano9']=cal.Ano1
cal.loc[cal.Mês.isin([1,2,3,4,5,6,7,8]),'Ano9']=cal.Ano1-1

cal['Ano10']=cal.Ano1
cal.loc[cal.Mês.isin([1,2,3,4,5,6,7,8,9]),'Ano10']=cal.Ano1-1

cal['Ano11']=cal.Ano1
cal.loc[cal.Mês.isin([1,2,3,4,5,6,7,8,9,10]),'Ano11']=cal.Ano1-1

cal['Ano12']=cal.Ano1
cal.loc[cal.Mês.isin([1,2,3,4,5,6,7,8,9,10,11]),'Ano12']=cal.Ano1-1

cal

r_pa=cal.merge(r_pa)
display(r_pa.head())

r_pa.to_csv('~/git/Importacao-por-SH-e-Municipio/temp/r_pa.csv',index=False)
