==> 7-Executa o teste completo.ipynb <==
%run '~/git/Bibliotecas.ipynb'

r_pa=pd.read_csv('~/git/Importacao-por-SH-e-Municipio/temp/r_pa.csv')
r_pa.head(1)

ano_tipo=['Ano1','Ano4','Ano7','Ano10']

quant=[ 'KT_f', 'miFOB_f', 'KT_c', 'miFOB_c','mm']

m_input=['rel_preço_ad','mm_ad','miFOB_f_ad', 'KT_c_ad', 'miFOB_c_ad']

import itertools as it

%%time
q=pd.DataFrame()

for a_t in ano_tipo:
    
    r_pa_at=r_pa[[a_t,'Mês']+quant].dropna()
    a_c=r_pa_at[[a_t,'Mês']].drop_duplicates()
    a_c=a_c.groupby(a_t)['Mês'].count().reset_index()
    a_c=a_c[a_c['Mês']==12][[a_t]]
    r_pa_at=r_pa_at.merge(a_c)
    r_pa_at=r_pa_at.groupby(a_t)[quant].sum().reset_index()
    r_pa_at['rel_preço']=(r_pa_at.miFOB_f/r_pa_at.KT_f)/(r_pa_at.miFOB_c/r_pa_at.KT_c)
    ad1=r_pa_at[[a_t,'KT_f']]
    ad2=r_pa_at.drop(columns=['KT_f'])
    ad2[a_t]=ad2[a_t]+1
    ad2=ad2.rename(columns={'miFOB_f':'miFOB_f_ad', 'KT_c':'KT_c_ad', 'miFOB_c':'miFOB_c_ad','rel_preço':'rel_preço_ad','mm':'mm_ad'})
    r_pa_ad_at=ad1.merge(ad2)

    
    a_i_a=r_pa_ad_at[a_t].min()
    
    anos_fim_treino=r_pa_ad_at[a_t].unique().tolist()
    
    for a_f_t in anos_fim_treino:
        
        for duração_treino in range(5,15,2):
            
            a_i_t=a_f_t-duração_treino
            
            if a_i_t>=a_i_a:

                r_pa_ad_at_train=r_pa_ad_at[
                    (r_pa_ad_at[a_t]<=a_f_t)
                    &
                    (r_pa_ad_at[a_t]>a_i_t)]

                train=r_pa_ad_at_train
                

                # Cria arquivos vazios para receber tests
                test_f=pd.DataFrame()
                #test_p=pd.DataFrame()

                # Se o ano para previsão f estiver contido em r_pa_ad_t
                if r_pa_ad_at_train[a_t].max()+1<=r_pa_ad_at[a_t].max():

                    # Define o test_f
                    test_f=r_pa_ad_at[r_pa_ad_at[a_t]==r_pa_ad_at_train[a_t].max()+1]


                if len(test_f)>0:


                    for n_comb in range(1,len(m_input)+1):

                        iterator=it.combinations(m_input,n_comb)

                        c_l=list(next(iterator))

                        while len(c_l)>0:

                            train_t=train[['KT_f']+c_l].dropna()

                            X_train=train_t.drop(columns='KT_f')
                            y_train=train_t.KT_f

                            LR.fit(X_train,y_train)
                            RFR.fit(X_train,y_train)

                            if len(test_f)>0:
                                test_f_t=test_f[['KT_f']+c_l]
                                X_test_f=test_f_t.drop(columns='KT_f')
                                y_test_f=test_f_t.KT_f.values[0]
                                test_f_LR=LR.predict(X_test_f)[0]
                                test_f_RFR=RFR.predict(X_test_f)[0]
                                
                            else:
                                y_test_f=np.nan
                                test_f_prev=np.nan


                            loop_result=pd.DataFrame({
                                'ano_tipo':[a_t],
                                'ano_ini_treino':[a_i_t],
                                'ano_fim_treino':[a_f_t],
                                'duração_treino':[duração_treino],
                                'c_l': ','.join(c_l),
                                'y_test_f':[y_test_f],
                                'test_f_LR':[test_f_LR],
                                'test_f_RFR':[test_f_RFR]
                                

                            })

                            q=pd.concat([q,loop_result])



                            try: c_l=list(next(iterator))
                            except: c_l=[]



display(q)

q['test_f%_LR'] =round(100*(q['test_f_LR' ]-q['y_test_f'])/q['y_test_f'])
q['test_f%_RFR']=round(100*(q['test_f_RFR']-q['y_test_f'])/q['y_test_f'])

q.to_csv('~/git/Importacao-por-SH-e-Municipio/temp/q1.csv',index=False)
