==> 8-Análise dos Testes Múltiplos-Copy1.ipynb <==
%run '~/git/Bibliotecas.ipynb'

r=pd.read_csv('~/git/Importacao-por-SH-e-Municipio/temp/q1.csv')
r

col=r.columns[[0,3,4]].tolist()
col

best=pd.DataFrame()
for c1 in col:
    a=np.abs(r.groupby(c1)[['test_f%_RFR']].mean()).rename(columns={'test_f%_RFR':'media_erro%'})

    b=np.abs(r.groupby(c1)[['test_f%_RFR']].std()).rename(columns={'test_f%_RFR':'std_erro%'})

    c=round(pd.concat([a,b],axis=1))

    c['erro_max']=c['media_erro%']+c['std_erro%']
    c=c.sort_values(by='erro_max').reset_index(drop=False)
    display(c)
    #display(c.loc[0,c1])
    best=pd.concat([best,c[[c1]].iloc[0]])
best=best.transpose()
display(best)

best.to_csv('~/git/Importacao-por-SH-e-Municipio/temp/best.csv',index=False)

r_b=r.merge(best)
r_b

r_b['test_f%_RFR'].mean()

r_b['test_f%_RFR'].std()
