==> 8-Análise dos Testes Múltiplos.ipynb <==
%run '~/git/Bibliotecas.ipynb'

r=pd.read_csv('~/git/Importacao-por-SH-e-Municipio/temp/q1.csv')

r['test_f%_LR'] =round(100*(r['test_f_LR' ]-r['y_test_f'])/r['y_test_f'])
r['test_f%_RFR']=round(100*(r['test_f_RFR']-r['y_test_f'])/r['y_test_f'])

a1=r.groupby('c_l')['test_f%_LR'].mean().reset_index().rename(columns={'test_f%_LR':'f%_LR_mean'})

a2=r.groupby('c_l')['test_f%_LR'].std().reset_index().rename(columns={'test_f%_LR':'f%_LR_std'})

a3=a1.merge(a2)
a3['desvio_maximo_LR']=np.abs(a3['f%_LR_mean'])+a3['f%_LR_std']

a1=r.groupby('c_l')['test_f%_RFR'].mean().reset_index().rename(columns={'test_f%_RFR':'f%_RFR_mean'})

a2=r.groupby('c_l')['test_f%_RFR'].std().reset_index().rename(columns={'test_f%_RFR':'f%_RFR_std'})

a3=a1.merge(a2)
a3['desvio_maximo_RFR']=np.abs(a3['f%_RFR_mean'])+a3['f%_RFR_std']

print(f"\n\nErro Médio das Regressões:\n\n{round(r[['test_f%_LR','test_f%_RFR']].mean())}")

print(f"\n\nDesvio Padrão das Regressões:\n\n{round(r[['test_f%_LR','test_f%_RFR']].std())}")

print('\n\nDevido a maior desvio padrão da Regressão Linear, seguiremos o estudo\n\
somente processando a regressão por Floresta Randômica.\n')

print('\n\n\nMelhores Combinações de Variáveis para Previsão\n')

a3=a3.sort_values(by=['desvio_maximo_RFR'])
display(round(a3).head(15).reset_index(drop=True))

Considerando todos os valores de desvio_maximo_FR sem diferença significativa, faremos opção por a combinação de fatores da linha 4, que alé do Ano1 , leava em conta também as chuvas e a relação de preço entre fertilizantes e café.

display(round(a3[a3.c_l=='Ano1']))

display(round(a3[a3.c_l=='Ano4']))

display(round(a3[a3.c_l=='Ano7']))

display(round(a3[a3.c_l=='Ano10']))

display(round(a3[a3.c_l=='mm_ad']))

display(round(a3[a3.c_l=='rel_preço_ad']))

display(round(a3[a3.c_l=='miFOB_f_ad']))

display(round(a3[a3.c_l=='miFOB_c_ad']))

display(round(a3[a3.c_l=='KT_c_ad']))

col_prev_multi_24=a3.iloc[4:5,:]
col_prev_multi_24.to_csv('~/git/Importacao-por-SH-e-Municipio/temp/col_prev_multi_24.csv')

print(f'\n\nPara a previsão do ano de 2024, utilizaremos a primeira combinação do rank acima,\n\
porém ressaltando que os 15 testes acima apresentaram resultados semelhantes,\n\n\
No teste feito para identificar o melhor mês para iniciar agrupamento de 12 meses\n\
se pode verificar que o agrupamento que apresentou a melhor performance foi o \n\
grupamento com início no mês 1, Janeiro, o que equivale ao ano comum, normalmente\n\
utilizado. Por isso, ele será utilizado na previsão para o ano 2024.\n\n\
A seguir, para completar a compreensão do tema, apresentamos o rank das\n\
piores combinações de características para a previsão.\n')

print('\nPiores Combinações de Variáveis para Previsão')
display(round(a3.sort_values(by=['desvio_maximo_RFR'], ascending=False).head(3)))

r.groupby('duração_treino')['test_f%_RFR'].mean().astype(int).reset_index().sort_values(by=['test_f%_RFR'])

print(f"Os testes objetivando identificar o melhor duração do período de treino para\n\
previsão não permitiram uma conclusão, como se pode ver abaixo:\n\n{r.groupby('duração_treino')['test_f%_RFR'].mean().astype(int).reset_index().sort_values(by=['test_f%_RFR'])}.\n\n\
Com isso, a previsão será feita utilizando a máxima duração possível.")

print('\n\nApós identificar qual combinção destes parâmetros proporcoina\n\
a maior precisão das previsões, faremos, em fim, a previsão para 2024.\n')
